fix(telemetry): address PR review feedback
- bot_manager: re-raise asyncio.CancelledError so graceful shutdown
  isn't blocked by the bot dispatch loop (qodo).
- bot.py: clarify the bot.run-before-active-check comment to match
  what the code actually does (copilot).
- test_bot_run_span: deterministic empty-message test using a
  whitespace-only template (copilot).
- test_metrics_webhook + test_webhook_http_span: deduplicate the
  webhook_server fixture into tests/telemetry/conftest.py
  (sonarcloud duplication gate); rename unused `server` to
  `_server` (sonar S1481); drop hand-rolled histogram-sum walk in
  favour of get_histogram_sum() helper (sonar S3776).
- _metrics_helpers: add get_histogram_sum() helper used by the
  duration-positive assertion.
