+++ tests/telemetry/test_metrics_webhook.py
@@ -3,39 +3,15 @@
 from __future__ import annotations
 
 import pytest
-import pytest_asyncio
 from aiohttp import ClientSession
 
-from culture.bots.bot_manager import BotManager
 from culture.bots.config import BotConfig
-from culture.bots.http_listener import HttpListener
-from tests.telemetry._metrics_helpers import get_histogram_count
-
-
-@pytest_asyncio.fixture
-async def webhook_server(server, tmp_path, monkeypatch):
-    """Server + listener on a random port; returns (server, mgr, port)."""
-    monkeypatch.setattr("culture.bots.config.BOTS_DIR", tmp_path)
-    monkeypatch.setattr("culture.bots.bot.BOTS_DIR", tmp_path)
-    monkeypatch.setattr("culture.bots.bot_manager.BOTS_DIR", tmp_path)
-
-    mgr = BotManager(server)
-    server.bot_manager = mgr
-
-    listener = HttpListener(mgr, "127.0.0.1", 0)
-    await listener.start()
-    site = next(iter(listener._runner._sites))
-    port = site._server.sockets[0].getsockname()[1]
-
-    yield server, mgr, port
-
-    await listener.stop()
-    await mgr.stop_all()
+from tests.telemetry._metrics_helpers import get_histogram_count, get_histogram_sum
 
 
 @pytest.mark.asyncio
 async def test_webhook_duration_records_2xx(metrics_reader, webhook_server):
-    server, mgr, port = webhook_server
+    _server, mgr, port = webhook_server
     await mgr.create_bot(
         BotConfig(
             name="testserv-hook",
@@ -116,7 +92,7 @@ async def test_webhook_duration_records_5xx_on_uncaught_exception(
     metrics_reader, webhook_server, monkeypatch
 ):
     """An uncaught non-ValueError/RuntimeError in dispatch is reported as 5xx."""
-    server, mgr, port = webhook_server
+    _server, mgr, port = webhook_server
     await mgr.create_bot(
         BotConfig(
             name="testserv-boom",
@@ -148,7 +124,7 @@ async def test_webhook_duration_records_5xx_on_uncaught_exception(
 @pytest.mark.asyncio
 async def test_webhook_duration_value_is_positive(metrics_reader, webhook_server):
     """Recorded duration must be > 0 (catches zero-time / inverted-timer regressions)."""
-    server, mgr, port = webhook_server
+    _server, mgr, port = webhook_server
     await mgr.create_bot(
         BotConfig(
             name="testserv-timed",
@@ -164,15 +140,9 @@ async def test_webhook_duration_value_is_positive(metrics_reader, webhook_server
         ) as resp:
             assert resp.status == 200
 
-    # Walk histogram points and assert sum > 0 for our bot.
-    data = metrics_reader.get_metrics_data()
-    total_sum = 0.0
-    for rm in data.resource_metrics:
-        for sm in rm.scope_metrics:
-            for metric in sm.metrics:
-                if metric.name != "culture.bot.webhook.duration":
-                    continue
-                for point in metric.data.data_points:
-                    if dict(point.attributes).get("bot") == "testserv-timed":
-                        total_sum += point.sum
-    assert total_sum > 0.0
+    total = get_histogram_sum(
+        metrics_reader,
+        "culture.bot.webhook.duration",
+        attrs={"bot": "testserv-timed"},
+    )
+    assert total > 0.0
