+++ tests/telemetry/test_webhook_http_span.py
@@ -3,39 +3,16 @@
 from __future__ import annotations
 
 import pytest
-import pytest_asyncio
 from aiohttp import ClientSession
 
-from culture.bots.bot_manager import BotManager
 from culture.bots.config import BotConfig
-from culture.bots.http_listener import HttpListener
-
-
-@pytest_asyncio.fixture
-async def webhook_server(server, tmp_path, monkeypatch):
-    monkeypatch.setattr("culture.bots.config.BOTS_DIR", tmp_path)
-    monkeypatch.setattr("culture.bots.bot.BOTS_DIR", tmp_path)
-    monkeypatch.setattr("culture.bots.bot_manager.BOTS_DIR", tmp_path)
-
-    mgr = BotManager(server)
-    server.bot_manager = mgr
-
-    listener = HttpListener(mgr, "127.0.0.1", 0)
-    await listener.start()
-    site = next(iter(listener._runner._sites))
-    port = site._server.sockets[0].getsockname()[1]
-
-    yield server, mgr, port
-
-    await listener.stop()
-    await mgr.stop_all()
 
 
 @pytest.mark.asyncio
 async def test_webhook_emits_aiohttp_server_span_parenting_bot_run(
     tracing_exporter, webhook_server
 ):
-    server, mgr, port = webhook_server
+    _server, mgr, port = webhook_server
     await mgr.create_bot(
         BotConfig(
             name="testserv-spanhook",
@@ -63,7 +40,6 @@ async def test_webhook_emits_aiohttp_server_span_parenting_bot_run(
     run_spans = [s for s in spans if s.name == "bot.run"]
     assert len(run_spans) == 1
     run = run_spans[0]
-    # bot.run is parented under the aiohttp server span (same trace).
     assert run.context.trace_id == aiohttp_span.context.trace_id
     assert run.parent is not None
     assert run.parent.span_id == aiohttp_span.context.span_id
